#31683 Merged updates from Pixomondo, cleanup, refactor
Merges in changes provided by Pixomondo (thanks guys!!). Now includes the
ability to define multiple output profiles in the configuration. Includes
methods for converting to and from regular, built-in Alembic nodes. Overall
cleanup and refactoring of the code to bring it more in line with other
toolkit apps (like nuke write node).

Closes #2.